+++ Sources/HostBox/IHosting.cs
@@ -1,0 +1,9 @@
+﻿using System.Threading.Tasks;
+
+namespace HostBox
+{
+    public interface IHosting
+    {
+        Task Run();
+    }
+}
